requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
pyyaml==6.0.1
openai==1.12.0
google-generativeai==0.3.2
jinja2==3.1.3
python-dotenv==1.0.1
schedule==1.2.1
